+++ imagickpixeliterator_class.c
@@ -466,7 +466,7 @@ PHP_METHOD(imagickpixeliterator, getpreviousiteratorrow)
 {
 	php_imagickpixeliterator_object *internpix;
 	PixelWand **wand_array;
-	unsigned long num_wands;
+	size_t num_wands;
 
 	if (zend_parse_parameters_none() == FAILURE) {
 		return;
@@ -504,7 +504,7 @@ PHP_METHOD(imagickpixeliterator, getcurrentiteratorrow)
 {
 	php_imagickpixeliterator_object *internpix;
 	PixelWand **wand_array;
-	unsigned long num_wands;
+	size_t num_wands;
 
 	if (zend_parse_parameters_none() == FAILURE) {
 		return;
@@ -545,7 +545,7 @@ PHP_METHOD(imagickpixeliterator, getnextiteratorrow)
 {
 	php_imagickpixeliterator_object *internpix;
 	PixelWand **wand_array;
-	unsigned long num_wands;
+	size_t num_wands;
 
 	if (zend_parse_parameters_none() == FAILURE) {
 		return;
